fix endregion syntax to allow region folding - refactor boots of comfort, bounty fist, buckler, carrot, cauldron stir, cloak, conjured ice pillar, cursed gear, and daggers
