+++ source/gear/_gearDictionary.js
@@ -17,30 +17,13 @@ for (const file of [
 	"battle-standard.js",
 	"blood-aegis.js",
 	"bonfire-formationjs",
-	"bootsofcomfort-accurate.js",
-	"bootsofcomfort-base.js",
-	"bootsofcomfort-hearty.js",
-	"bootsofcomfort-powerful.js",
-	"bountyfist-base.js",
-	"bountyfist-midass.js",
-	"bountyfist-thirsting.js",
-	"buckler-base.js",
-	"buckler-guarding.js",
-	"buckler-supportive.js",
-	"carrot-balanced.js",
-	"carrot-base.js",
-	"carrot-guarding.js",
-	"cauldronstir-base.js",
-	"cauldronstir-incompatible.js",
-	"cauldronstir-innovative.js",
-	"cauldronstir-sabotaging.js",
-	"cauldronstir-toxic.js",
-	"cloak-accurate.js",
-	"cloak-base.js",
-	"cloak-powerful.js",
-	"conjuredicepillar-base.js",
-	"conjuredicepillar-taunting.js",
-	"conjuredicepillar-devoted.js",
+	"boots-of-comfort.js",
+	"bounty-fist.js",
+	"buckler.js",
+	"carrot.js",
+	"cauldron-stir.js",
+	"cloak.js",
+	"conjured-ice-pillar.js",
 	"cursed-bag.js",
 	"cursed-blade.js",
 	"cursed-doll.js",
@@ -48,11 +31,7 @@ for (const file of [
 	"cursed-scroll.js",
 	"cursed-shield.js",
 	"cursed-tome.js",
-	"daggers-attuned",
-	"daggers-balanced",
-	"daggers-base",
-	"daggers-centering",
-	"daggers-distracting",
+	"daggers.js",
 	"deckofcards-base.js",
 	"deckofcards-evasive.js",
 	"deckofcards-numbing.js",
