--- source/gear/bountyfist-midass.js
@@ -1,35 +1,0 @@
-const { GearTemplate } = require('../classes');
-const { ESSENCE_MATCH_STAGGER_FOE } = require('../constants');
-const { changeStagger, dealDamage, generateModifierResultLines, addModifier } = require('../util/combatantUtil');
-const { damageScalingGenerator } = require('./shared/scalings');
-
-const variantName = "Midas's Bounty Fist";
-module.exports = new GearTemplate(variantName,
-	[
-		["use", "Inflict <@{damage} + gold paid> @{essence} damage and @{mod0Stacks} @{mod0} on a foe"],
-		["critical", "Damage x @{critBonus}"]
-	],
-	"Pact",
-	"Earth"
-).setCost(350)
-	.setEffect((targets, user, adventure) => {
-		const { essence, pactCost, scalings: { damage, critBonus }, modifiers: [curseOfMidas] } = module.exports;
-		const goldUsed = Math.floor(adventure.gold * (pactCost[0] / 100));
-		adventure.gold -= goldUsed;
-		let pendingDamage = damage.calculate(user) + goldUsed;
-		if (user.crit) {
-			pendingDamage *= critBonus;
-		}
-		const { resultLines, survivors } = dealDamage(targets, user, pendingDamage, false, essence, adventure);
-		if (user.essence === essence) {
-			changeStagger(survivors, user, ESSENCE_MATCH_STAGGER_FOE);
-		}
-		return resultLines.concat(generateModifierResultLines(addModifier(targets, curseOfMidas)), `${user.name}'s ${variantName} consumed ${goldUsed}g.`);
-	}, { type: "single", team: "foe" })
-	.setSidegrades("Thirsting Bounty Fist")
-	.setPactCost([10, "@{pactCost}% of party gold"])
-	.setScalings({
-		damage: damageScalingGenerator(40),
-		critBonus: 2
-	})
-	.setModifiers({ name: "Curse of Midas", stacks: 2 });
