+++ source/gear/cursed-doll.js
@@ -1,6 +1,6 @@
-const { GearTemplate } = require('../classes');
+const { GearTemplate, GearFamily } = require('../classes');
 
-module.exports = new GearTemplate("Cursed Doll",
+const cursedDoll = new GearTemplate("Cursed Doll",
 	[
 		["Passive", "Reduces your Crit Rate by @{percentCritRate*-1}%, can be upgraded into random Support gear"],
 	],
@@ -9,3 +9,5 @@ module.exports = new GearTemplate("Cursed Doll",
 ).setCost(-50)
 	.setUpgrades("Arcane Sledge", "Carrot", "Medic's Kit", "Elemental Scroll", "Midas Staff", "War Cry")
 	.setScalings({ percentCritRate: -5 });
+
+module.exports = new GearFamily(cursedDoll, [], true);
