+++ source/gear/cursed-grimore.js
@@ -1,6 +1,6 @@
-const { GearTemplate } = require('../classes');
+const { GearTemplate, GearFamily } = require('../classes');
 
-module.exports = new GearTemplate("Cursed Grimore",
+const cursedGrimore = new GearTemplate("Cursed Grimore",
 	[
 		["Passive", "Can be upgraded into a random Pact"],
 	],
@@ -8,3 +8,5 @@ module.exports = new GearTemplate("Cursed Grimore",
 	"Unaligned"
 ).setCost(-50)
 	.setUpgrades("Tempestuous Wrath", "Bounty Fist", "Universal Solution", "Overburn Explosion", "Forbidden Knowledge", "Blood Aegis");
+
+module.exports = new GearFamily(cursedGrimore, [], true);
