+++ source/gear/cursed-shield.js
@@ -1,6 +1,6 @@
-const { GearTemplate } = require('../classes');
+const { GearTemplate, GearFamily } = require('../classes');
 
-module.exports = new GearTemplate("Cursed Shield",
+const cursedShield = new GearTemplate("Cursed Shield",
 	[
 		["Passive", "Reduces your Power by @{percentPower*-1}%, can upgrade into random Defense gear"],
 	],
@@ -9,3 +9,5 @@ module.exports = new GearTemplate("Cursed Shield",
 ).setCost(-50)
 	.setUpgrades("Enchantment Siphon", "Smokescreen", "Buckler", "Cloak", "Parrying Dagger", "Spiked Shield")
 	.setScalings({ percentPower: -5 });
+
+module.exports = new GearFamily(cursedShield, [], true);
