--- source/gear/daggers-attuned.js
@@ -1,29 +1,0 @@
-const { GearTemplate } = require('../classes');
-const { ESSENCE_MATCH_STAGGER_FOE } = require('../constants');
-const { dealDamage, changeStagger, generateModifierResultLines, addModifier, combineModifierReceipts } = require('../util/combatantUtil');
-const { archetypeActionDamageScaling } = require('./shared/scalings');
-
-module.exports = new GearTemplate("Attuned Daggers",
-	[
-		["use", "Deal <@{damage}> @{essence} damage to a foe and gain @{mod0Stacks} @{mod0} and @{mod1Stacks} @{mod1}"],
-		["critical", "Damage x @{critBonus}"]
-	],
-	"Action",
-	"Fire"
-).setEffect((targets, user, adventure) => {
-	const { essence, scalings: { damage, critBonus }, modifiers: [excellence, attunement] } = module.exports;
-	let pendingDamge = damage.calculate(user);
-	if (user.crit) {
-		pendingDamge *= critBonus;
-	}
-	const { resultLines, survivors } = dealDamage(targets, user, pendingDamge, false, essence, adventure);
-	if (user.essence === essence) {
-		changeStagger(survivors, user, ESSENCE_MATCH_STAGGER_FOE);
-	}
-	return resultLines.concat(generateModifierResultLines(combineModifierReceipts(addModifier([user], excellence).concat(addModifier([user], attunement)))));
-}, { type: "single", team: "foe" })
-	.setScalings({
-		damage: archetypeActionDamageScaling,
-		critBonus: 2
-	})
-	.setModifiers({ name: "Excellence", stacks: 2 }, { name: "Attunement", stacks: 2 });
